Cleanup LASSI unittests
