+++ tests/lassi/test_c2h4n4.py
@@ -134,56 +134,6 @@ class KnownValues(unittest.TestCase):
         # 2 + 3 + 3 + 3 + 2 = 13
         self.assertEqual (las3.nroots, 13)
 
-    def test_casci_limit (self):
-        from mrh.my_pyscf.lassi.states import all_single_excitations
-        from mrh.my_pyscf.mcscf.lasci import get_space_info
-        xyz='''H 0 0 0
-        H 1 0 0
-        H 3 0 0
-        H 4 0 0'''
-        rmol = gto.M (atom=xyz, basis='sto3g', symmetry=False, verbose=0, output='/dev/null')
-        rmf = scf.RHF (rmol).run ()
-
-        # Random Hamiltonian
-        rng = np.random.default_rng ()
-        rmf._eri = rng.random (rmf._eri.shape)
-        hcore = rng.random ((4,4))
-        rmf.get_hcore = lambda *args: hcore
-
-        # CASCI limit
-        mc = mcscf.CASCI (rmf, 4, 4).run ()
-        casdm1, casdm2 = mc.fcisolver.make_rdm12 (mc.ci, mc.ncas, mc.nelecas)
-
-        # LASSCF
-        rlas = LASSCF (rmf, (2,2), (2,2), spin_sub=(1,1))
-        rlas.conv_tol_grad = rlas.conv_tol_self = 9e99
-
-        # LASSI in the CASCI limit
-        for i in range (2): rlas = all_single_excitations (rlas)
-        charges, spins, smults, wfnsyms = get_space_info (rlas)
-        lroots = 4 - smults
-        idx = (charges!=0) & (lroots==3)
-        lroots[idx] = 1
-        rlas.lasci (lroots=lroots.T)
-        e_roots, si = rlas.lassi (opt=0)
-        with self.subTest ("total energy"):
-            self.assertAlmostEqual (e_roots[0], mc.e_tot, 8)
-        lasdm1s, lasdm2s = root_make_rdm12s (rlas, rlas.ci, si, state=0, opt=0)
-        lasdm1 = lasdm1s.sum (0)
-        lasdm2 = lasdm2s.sum ((0,3))
-        with self.subTest ("casdm1"):
-            self.assertAlmostEqual (lib.fp (lasdm1), lib.fp (casdm1), 8)
-        with self.subTest ("casdm2"):
-            self.assertAlmostEqual (lib.fp (lasdm2), lib.fp (casdm2), 8)
-        stdm1s = make_stdm12s (rlas, opt=0)[0][9:13,:,:,:,9:13] # second rootspace
-        with self.subTest("state indexing"):
-            # column-major ordering for state excitation quantum numbers:
-            # earlier fragments advance faster than later fragments
-            self.assertAlmostEqual (lib.fp (stdm1s[0,:,:2,:2,0]),
-                                    lib.fp (stdm1s[2,:,:2,:2,2]))
-            self.assertAlmostEqual (lib.fp (stdm1s[0,:,2:,2:,0]),
-                                    lib.fp (stdm1s[1,:,2:,2:,1]))
-
 if __name__ == "__main__":
     print("Full Tests for SA-LASSI of c2h4n4 molecule")
     unittest.main()
